tmux: fix split directions
Apparently tmux made the very confusing decision to confuse their own
split definitions. They list % as a "vertical" split and " as a
"horizontal" split, but just doing it you can tell that is wrong. You
need to rebind the keys with the opposite flags. Or maybe every
documenting page online is wrong. Idk, this is stupid. Bah.